requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
appdirs==1.4.3            # via quilt
attrs==18.2.0             # via fiona
backcall==0.1.0           # via ipython
bleach==3.0.2             # via nbconvert
certifi==2018.8.24
chardet==3.0.4            # via requests
click-plugins==1.0.4      # via fiona
click==7.0                # via click-plugins, cligj, fiona
cligj==0.5.0              # via fiona
colorlover==0.2.1         # via cufflinks
cufflinks==0.14.6
cycler==0.10.0            # via matplotlib
decorator==4.3.0          # via ipython, plotly, traitlets
defusedxml==0.5.0         # via nbconvert
entrypoints==0.2.3        # via nbconvert
fiona==1.8.4              # via geopandas
future==0.17.1            # via quilt
geopandas==0.4.0
idna==2.8                 # via requests
ipykernel==5.1.0          # via ipywidgets, notebook
ipython-genutils==0.2.0   # via nbformat, notebook, traitlets
ipython==7.2.0            # via cufflinks, ipykernel, ipywidgets, vdom
ipywidgets==7.4.2
jedi==0.13.2              # via ipython
jinja2==2.10              # via nbconvert, notebook
jsonschema==2.6.0         # via jupyterlab-launcher, jupyterlab-server, nbformat, vdom
jupyter-client==5.2.4     # via ipykernel, notebook
jupyter-core==4.4.0       # via jupyter-client, nbconvert, nbformat, notebook
jupyterlab-launcher==0.13.1
jupyterlab-server==0.2.0  # via jupyterlab
jupyterlab==0.35.4
kiwisolver==1.0.1         # via matplotlib
markupsafe==1.1.0         # via jinja2
matplotlib==3.0.2
mistune==0.8.4            # via nbconvert
munch==2.3.2              # via fiona
nbconvert==5.4.0          # via notebook
nbformat==4.4.0           # via ipywidgets, nbconvert, notebook, plotly
nose==1.3.7               # via cufflinks
notebook==5.7.3           # via jupyterlab, jupyterlab-launcher, jupyterlab-server, widgetsnbextension
numpy==1.15.4
packaging==18.0           # via quilt
pandas==0.23.4
pandocfilters==1.4.2      # via nbconvert
parso==0.3.1              # via jedi
pexpect==4.6.0            # via ipython
pickleshare==0.7.5        # via ipython
pint==0.8.1
plotly==3.4.2
prometheus-client==0.5.0  # via notebook
prompt-toolkit==2.0.7     # via ipython
ptyprocess==0.6.0         # via pexpect, terminado
pyarrow==0.11.1           # via quilt
pygments==2.3.1           # via ipython, nbconvert
pyparsing==2.3.0          # via matplotlib, packaging
pyproj==1.9.5.1           # via geopandas
python-dateutil==2.7.5
pytz==2018.7              # via pandas, plotly
pyyaml==3.13              # via quilt
pyzmq==17.1.2             # via jupyter-client, notebook
quilt==2.9.13
requests==2.21.0          # via plotly, quilt
retrying==1.3.3           # via plotly
scipy==1.1.0
send2trash==1.5.0         # via notebook
shapely==1.6.4.post2      # via geopandas
six==1.12.0               # via bleach, cufflinks, cycler, fiona, munch, packaging, plotly, prompt-toolkit, pyarrow, python-dateutil, quilt, retrying, traitlets
terminado==0.8.1          # via notebook
testpath==0.4.2           # via nbconvert
tornado==5.1.1            # via ipykernel, jupyter-client, notebook, terminado
tqdm==4.28.1              # via quilt
traitlets==4.3.2
urllib3==1.24.1           # via requests
vdom==0.6
wcwidth==0.1.7            # via prompt-toolkit
webencodings==0.5.1       # via bleach
widgetsnbextension==3.4.2  # via ipywidgets
xlrd==1.2.0               # via quilt
